+++ packages/cadl-ranch/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @azure-tools/cadl-ranch
 
+## 0.2.8
+
+### Patch Changes
+
+- [#149](https://github.com/Azure/cadl-ranch/pull/149) [`66814a8`](https://github.com/Azure/cadl-ranch/commit/66814a8ae6c4fe2f8bc61e8fc8703f0a74f0bf98) Thanks [@changlong-liu](https://github.com/changlong-liu)! - - add testcases for using special words as operation/parameter/property names
+
+## 0.2.7
+
+### Patch Changes
+
+- [#142](https://github.com/Azure/cadl-ranch/pull/142) [`9116e47`](https://github.com/Azure/cadl-ranch/commit/9116e478df9cba0aad2bee1d23ed59382260c325) Thanks [@changlong-liu](https://github.com/changlong-liu)! - fix lro mockapi: add response body for Azure_Lro_PollingSuccess
+
 ## 0.2.6
 
 ### Patch Changes